requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8

Pillow==6.2.1

PySocks==1.7.1

six==1.13.0

urllib3==1.25.7
